chore: Refactor EchoMaker form validation and component structure
