--- src/components/FileUpload.jsx
@@ -1,10 +1,0 @@
-import React, { useState } from "react";
-
-export default function FileUpload({ handleChange }) {
-  return (
-    <div>
-      <label htmlFor="media">Select Media: </label>
-      <input type="file" name="media" onChange={handleChange} multiple />
-    </div>
-  );
-}
